Separated user and expert flow. Implemented state persist for message
