+++ byoeb-v1/byoeb/byoeb/services/chat/constants.py
@@ -1,0 +1,32 @@
+EMOJI = "emoji"
+VERIFICATION_STATUS = "verification_status"
+
+DESCRIPTION = "description"
+ROW_TEXTS = "row_texts"
+RELATED_QUESTIONS = "related_questions"
+
+DATA = "data"
+MIME_TYPE = "mime_type"
+
+USER = "user"
+MESSAGES_CONTEXT = "messages_context"
+
+TEMPLATE_NAME = "template_name"
+TEMPLATE_LANGUAGE = "template_language"
+TEMPLATE_PARAMETERS = "template_parameters"
+
+CREATE = "create"
+UPDATE = "update"
+
+PENDING = "pending"
+WAITING = "waiting"
+WRONG = "wrong"
+VERIFIED = "verified"
+
+CORRECTED_ANSWER_EN = "corrected_answer_en"
+CORRECTED_ANSWER_SOURCE = "corrected_answer_source"
+
+CORRECTION_EN = "correction_en"
+CORRECTION_SOURCE = "correction_source"
+
+UPDATE_ID = "update_id"
