+++ byoeb-v1/byoeb/byoeb/services/chat/message_handlers/__init__.py
@@ -1,6 +1,5 @@
-from byoeb.services.chat.message_handlers.process import ByoebUserProcess, ByoebExpertProcess
-from byoeb.services.chat.message_handlers.generate import ByoebUserGenerateResponse, ByoebExpertGenerateResponse
-from byoeb.services.chat.message_handlers.send import ByoebUserSendResponse, ByoebExpertSendResponse
+from byoeb.services.chat.message_handlers.user_flow_handlers import ByoebUserProcess, ByoebUserGenerateResponse, ByoebUserSendResponse
+from byoeb.services.chat.message_handlers.expert_flow_handlers import ByoebExpertProcess, ByoebExpertGenerateResponse, ByoebExpertSendResponse
 
 __all__ = [
     "ByoebUserProcess",
