+++ byoeb-v1/byoeb/byoeb/services/chat/message_handlers/user_flow_handlers/process.py
@@ -46,21 +46,3 @@ class ByoebUserProcess(Handler):
         
         if self._successor:
             return await self._successor.handle([message])
-
-class ByoebExpertProcess(Handler):
-    
-    async def handle(
-        self,
-        messages: List[ByoebMessageContext]
-    ):
-        from byoeb.app.configuration.dependency_setup import text_translator
-        message = messages[0]
-        translated_en_text = await text_translator.atranslate_text(
-            input_text=message.message_context.message_source_text,
-            source_language=message.user.user_language,
-            target_language="en"
-        )
-        message.message_context.message_english_text = translated_en_text
-        if self._successor:
-            return await self._successor.handle([message])
-
